+++ src/coupon/domain/dto/coupon.result.dto.ts
