+++ src/coupon/presentation/dto/get-all-coupons.res.dto.ts
@@ -1,4 +1,4 @@
-import { CouponResult } from "@app/coupon/domain/dto/coupon.result";
+import { CouponResult } from "@app/coupon/domain/dto/coupon.result.dto";
 import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
 import { IsArray } from "class-validator";
