+++ src/coupon/presentation/dto/get-coupons-by-member.res.dto.ts
@@ -1,4 +1,4 @@
-import { MemberCouponResult } from "@app/coupon/domain/dto/member_coupon.result";
+import { MemberCouponResult } from "@app/coupon/domain/dto/member_coupon.result.dto";
 import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
 import { IsArray } from "class-validator";
