+++ src/member/domain/service/member.service.spec.ts
@@ -1,11 +1,11 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { MemberService } from "./member.service";
-import { GetBalanceCommand } from "../dto/get-balance.command";
-import { ChargeBalanceCommand } from "../dto/charge-balance.command";
-import { UseBalanceCommand } from "../dto/use-balance.command";
 import { MemberRepository } from "../../infrastructure/memeber.repository";
+import { GetBalanceCommand } from "../dto/get-balance.command.dto";
+import { ChargeBalanceCommand } from "../dto/charge-balance.command.dto";
+import { UseBalanceCommand } from "../dto/use-balance.command.dto";
 import { Member } from "../entity/member.entity";
-import { BalanceResult } from "../dto/balance.result";
+import { BalanceResult } from "../dto/balance.result.dto";
 import { BalanceHisotryRepository } from "../../infrastructure/balanceHistory.repository";
 import { IMEMBER_REPOSITORY } from "../../repository/member.repository.interface";
 import { IBALANCE_HISTORY_REPOSITORY } from "../../repository/balanceHistory.repository.interface";
