+++ src/order/order.module.ts
@@ -2,10 +2,10 @@ import { Module } from "@nestjs/common";
 import { OrderService } from "./domain/service/order.service";
 import { OrderController } from "./presentation/order.controller";
 import { PrismaModule } from "@app/database/prisma/prisma.module";
-import { OrderRepository } from "./domain/infrastructure/order.repository";
-import { OrderProductRepository } from "./domain/infrastructure/order_product.repository";
-import { IORDER_REPOSITORY } from "./domain/order.repository.interface";
-import { IORDER_PRODUCT_REPOSITORY } from "./domain/order_product.repository.interface";
+import { OrderRepository } from "./infrastructure/order.repository";
+import { OrderProductRepository } from "./infrastructure/order_product.repository";
+import { IORDER_REPOSITORY } from "./repository/order.repository.interface";
+import { IORDER_PRODUCT_REPOSITORY } from "./repository/order_product.repository.interface";
 
 @Module({
   imports: [PrismaModule],
