+++ src/payment/domain/service/payment.service.spec.ts
@@ -2,8 +2,8 @@ import { Test, TestingModule } from "@nestjs/testing";
 import { PaymentService } from "./payment.service";
 import { IPAYMENT_REPOSITORY } from "../../repository/payment.repository.interface";
 import { PaymentStatus } from "../dto/payment-status.enum";
-import { ProcessPaymentCommand } from "../dto/process-payment.command";
-import { PaymentResult } from "../dto/payment.result";
+import { ProcessPaymentCommand } from "../dto/process-payment.command.dto";
+import { PaymentResult } from "../dto/payment.result.dto";
 
 describe("PaymentService", () => {
   let paymentService: PaymentService;
