+++ src/payment/payment.module.ts
@@ -2,8 +2,8 @@ import { PrismaModule } from "@app/database/prisma/prisma.module";
 import { Module } from "@nestjs/common";
 import { PaymentController } from "./presentation/payment.controller";
 import { PaymentService } from "./domain/service/payment.service";
-import { PaymentRepository } from "./domain/infrastructure/payment.repository";
-import { IPAYMENT_REPOSITORY } from "./domain/payment.repository.interface";
+import { PaymentRepository } from "./infrastructure/payment.repository";
+import { IPAYMENT_REPOSITORY } from "./repository/payment.repository.interface";
 import { MemberModule } from "@app/member/member.module";
 import { OrderModule } from "@app/order/order.module";
 import { ProductModule } from "@app/product/product.module";
