+++ src/product/domain/dto/add-stock.command.dto.ts
