+++ src/product/domain/dto/deduct-stock.command.dto.ts
