+++ src/product/domain/dto/get-product.command.dto.ts
