+++ src/product/domain/dto/product.result.dto.ts
