+++ src/product/domain/service/product.service.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from "@nestjs/common";
-import { ProductResult } from "../dto/product.result";
-import { GetProductCommand } from "../dto/get-product.command";
-import { AddStockCommand } from "../dto/add-stock.command";
-import { DeductStockCommand } from "../dto/deduct-stock.command";
+import { ProductResult } from "../dto/product.result.dto";
+import { GetProductCommand } from "../dto/get-product.command.dto";
+import { AddStockCommand } from "../dto/add-stock.command.dto";
+import { DeductStockCommand } from "../dto/deduct-stock.command.dto";
 import { ProductRepository } from "../../infrastructure/product.repository";
 import { Product } from "@prisma/client";
 import { IPRODUCT_REPOSITORY } from "../../repository/product.repository.interface";
