+++ src/productSalesStat/domain/dto/add-product-sales-stat.command.dto.ts
