+++ src/productSalesStat/domain/dto/add-product-sales-stat.result.dto.ts
