+++ src/productSalesStat/domain/dto/product-sales-stat.result.dto.ts
