+++ src/productSalesStat/domain/service/producSalesStat.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { ProductSalesStatRepository } from "../../infrastructure/product_sales_stat.repository";
 import { ProductSalesStatService } from "./productSalesStat.service";
-import { AddProductSalesStatCommand, PaidProduct } from "../dto/add-product-sales-stat.command";
+import { AddProductSalesStatCommand, PaidProduct } from "../dto/add-product-sales-stat.command.dto";
 import { IPRODUCT_SALES_STAT_REPOSITORY } from "../../repository/product_sales_stat.interface.repository";
 
 describe("ProductSalesStatService", () => {
