+++ src/productSalesStat/domain/service/productSalesStat.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable } from "@nestjs/common";
-import { ProductSalesStatResult } from "../dto/product-sales-stat.result";
-import { AddProductSalesStatResult } from "../dto/add-product-sales-stat.result";
+import { ProductSalesStatResult } from "../dto/product-sales-stat.result.dto";
+import { AddProductSalesStatResult } from "../dto/add-product-sales-stat.result.dto";
 import { Product_Sales_Stat } from "@prisma/client";
-import { AddProductSalesStatCommand, PaidProduct } from "../dto/add-product-sales-stat.command";
 import { ProductSalesStatRepository } from "../../infrastructure/product_sales_stat.repository";
+import { AddProductSalesStatCommand, PaidProduct } from "../dto/add-product-sales-stat.command.dto";
 import { IPRODUCT_SALES_STAT_REPOSITORY } from "../../repository/product_sales_stat.interface.repository";
 
 @Injectable()
