+++ src/productSalesStat/presentation/dto/get-popular-products.res.dto.ts
@@ -1,4 +1,4 @@
-import { ProductSalesStatResult } from "../../domain/dto/product-sales-stat.result";
+import { ProductSalesStatResult } from "../../domain/dto/product-sales-stat.result.dto";
 import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
 import { IsArray } from "class-validator";
