+++ src/productSalesStat/productSalesStat.module.ts
@@ -2,8 +2,8 @@ import { PrismaModule } from "@app/database/prisma/prisma.module";
 import { Module } from "@nestjs/common";
 import { ProductSalesStatController } from "./presentation/productSalesStat.controller";
 import { ProductSalesStatService } from "./domain/service/productSalesStat.service";
-import { ProductSalesStatRepository } from "./domain/infrastructure/product_sales_stat.repository";
-import { IPRODUCT_SALES_STAT_REPOSITORY } from "./domain/product_sales_stat.interface.repository";
+import { ProductSalesStatRepository } from "./infrastructure/product_sales_stat.repository";
+import { IPRODUCT_SALES_STAT_REPOSITORY } from "./repository/product_sales_stat.interface.repository";
 
 @Module({
   imports: [PrismaModule],
